+++ KG_Task_3_Math/src/com/cg/vsu/vectors/Vector2f.java
@@ -35,13 +35,13 @@ public class Vector2f implements Vector<Vector2f> {
     }
 
     @Override
-    public double length() {
-        return Math.sqrt(Math.pow(x, 2) + Math.pow(y, 2));
+    public float length() {
+        return (float) Math.sqrt(Math.pow(x, 2) + Math.pow(y, 2));
     }
 
     @Override
     public void normalize() {
-        float length = (float) this.length();
+        float length = this.length();
         x = x / length;
         y = y / length;
     }
