+++ KG_Task_3_Math/src/com/cg/vsu/vectors/Vector4f.java
@@ -41,13 +41,13 @@ public class Vector4f implements Vector<Vector4f> {
     }
 
     @Override
-    public double length() {
-        return Math.sqrt(Math.pow(x, 2) + Math.pow(y, 2) + Math.pow(z, 2) + Math.pow(w, 2));
+    public float length() {
+        return (float) Math.sqrt(Math.pow(x, 2) + Math.pow(y, 2) + Math.pow(z, 2) + Math.pow(w, 2));
     }
 
     @Override
     public void normalize() {
-        float length = (float) this.length();
+        float length = this.length();
         x = x / length;
         y = y / length;
         z = z / length;
